Add files via upload
This is a tableau dashboard and its twb file which shows how a persons parameter like CGPA, Toefl, Gre, research publication and other parameters effect the admit chances of a person. It also displays the average for each parameter for the given university dataset and allows filtering option to focus on whichever facet needed for analyzation.