Add files via upload
This file we have analyzed the parameters on which university decide admits using Tableau. We have created an interactive dashboard which showcases all data and uses filtering and sorting options. We also use various graphs and highlighters to denote the data and also calculate the averages of the parameters and its percentage of admit conversion.